Add connection info widgets and clean up device dialog
