+++ packages/devtools_app/lib/src/shared/ui/vm_flag_widgets.dart
@@ -13,8 +13,14 @@ import '../analytics/analytics.dart' as ga;
 import '../analytics/constants.dart' as gac;
 import '../banner_messages.dart';
 import '../common_widgets.dart';
+import '../dialogs.dart';
 import '../globals.dart';
+import '../primitives/auto_dispose.dart';
+import '../primitives/utils.dart';
+import '../table/table.dart';
+import '../table/table_data.dart';
 import '../theme.dart';
+import '../utils.dart';
 
 /// DropdownButton that controls the value of the 'profile_period' vm flag.
 ///
@@ -101,4 +107,205 @@ class CpuSamplingRateDropdown extends StatelessWidget {
       newValue ?? mediumProfilePeriod,
     );
   }
+}
+
+class ViewVmFlagsButton extends StatelessWidget {
+  const ViewVmFlagsButton({
+    super.key,
+    required this.gaScreen,
+    this.elevated = false,
+    this.minScreenWidthForTextBeforeScaling,
+  });
+
+  final String gaScreen;
+
+  final bool elevated;
+
+  final double? minScreenWidthForTextBeforeScaling;
+
+  @override
+  Widget build(BuildContext context) {
+    return DevToolsButton(
+      elevatedButton: elevated,
+      label: 'View VM flags',
+      icon: Icons.flag_rounded,
+      gaScreen: gaScreen,
+      gaSelection: gac.HomeScreenEvents.viewVmFlags.name,
+      minScreenWidthForTextBeforeScaling: minScreenWidthForTextBeforeScaling,
+      onPressed: () {
+        Navigator.of(context, rootNavigator: true).pop('dialog');
+        unawaited(
+          showDialog(
+            context: context,
+            builder: (context) => const VMFlagsDialog(),
+          ),
+        );
+      },
+    );
+  }
+}
+
+class VMFlagsDialog extends StatefulWidget {
+  const VMFlagsDialog({super.key});
+
+  @override
+  State<VMFlagsDialog> createState() => _VMFlagsDialogState();
+}
+
+class _VMFlagsDialogState extends State<VMFlagsDialog> with AutoDisposeMixin {
+  late final TextEditingController filterController;
+
+  var flags = <_DialogFlag>[];
+
+  var filteredFlags = <_DialogFlag>[];
+
+  @override
+  void initState() {
+    super.initState();
+
+    filterController = TextEditingController();
+    addAutoDisposeListener(filterController, () {
+      setState(() {
+        _refilter();
+      });
+    });
+
+    _updateFromController();
+    addAutoDisposeListener(serviceManager.vmFlagManager.flags, () {
+      setState(() {
+        _updateFromController();
+      });
+    });
+  }
+
+  void _updateFromController() {
+    flags = (serviceManager.vmFlagManager.flags.value?.flags ?? [])
+        .map((flag) => _DialogFlag(flag))
+        .toList();
+    _refilter();
+  }
+
+  void _refilter() {
+    final filter = filterController.text.trim().toLowerCase();
+
+    filteredFlags = filter.isEmpty
+        ? flags
+        : flags.where((flag) => flag.filterText.contains(filter)).toList();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return DevToolsDialog(
+      title: Row(
+        children: [
+          const DialogTitleText('VM Flags'),
+          const Expanded(child: SizedBox(width: denseSpacing)),
+          SizedBox(
+            width: defaultSearchFieldWidth,
+            height: defaultTextFieldHeight,
+            child: TextField(
+              controller: filterController,
+              decoration: const InputDecoration(
+                isDense: true,
+                border: OutlineInputBorder(),
+                labelText: 'Filter',
+              ),
+            ),
+          ),
+        ],
+      ),
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          SizedBox(
+            width: 875,
+            height: 375,
+            child: _FlagTable(filteredFlags),
+          ),
+        ],
+      ),
+      actions: const [
+        DialogCloseButton(),
+      ],
+    );
+  }
+}
+
+class _FlagTable extends StatelessWidget {
+  const _FlagTable(this.flags);
+
+  final List<_DialogFlag> flags;
+
+  static final name = _NameColumn();
+  static final description = _DescriptionColumn();
+  static final value = _ValueColumn();
+  static final columns = <ColumnData<_DialogFlag>>[name, description, value];
+
+  @override
+  Widget build(BuildContext context) {
+    return OutlineDecoration(
+      child: FlatTable<_DialogFlag>(
+        keyFactory: (_DialogFlag flag) => ValueKey<String?>(flag.name),
+        data: flags,
+        dataKey: 'vm-flags',
+        columns: columns,
+        defaultSortColumn: name,
+        defaultSortDirection: SortDirection.ascending,
+      ),
+    );
+  }
+}
+
+class _NameColumn extends ColumnData<_DialogFlag> {
+  _NameColumn()
+      : super(
+          'Name',
+          fixedWidthPx: scaleByFontFactor(180),
+        );
+
+  @override
+  String getValue(_DialogFlag dataObject) => dataObject.name ?? '';
+}
+
+class _DescriptionColumn extends ColumnData<_DialogFlag> {
+  _DescriptionColumn()
+      : super.wide(
+          'Description',
+          minWidthPx: scaleByFontFactor(100),
+        );
+
+  @override
+  String getValue(_DialogFlag dataObject) => dataObject.description ?? '';
+
+  @override
+  String getTooltip(_DialogFlag dataObject) => getValue(dataObject);
+}
+
+class _ValueColumn extends ColumnData<_DialogFlag> {
+  _ValueColumn()
+      : super(
+          'Value',
+          fixedWidthPx: scaleByFontFactor(160),
+          alignment: ColumnAlignment.right,
+        );
+
+  @override
+  String getValue(_DialogFlag dataObject) => dataObject.value ?? '';
+}
+
+class _DialogFlag {
+  _DialogFlag(this.flag)
+      : filterText = '${flag.name?.toLowerCase()}\n'
+            '${flag.comment?.toLowerCase()}\n'
+            '${flag.valueAsString?.toLowerCase()}';
+
+  final Flag flag;
+  final String filterText;
+
+  String? get name => flag.name;
+
+  String? get description => flag.comment;
+
+  String? get value => flag.valueAsString;
 }
